+++ app/ios/Runner/GoogleService-Info.plist
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>911090474857-gjmq3t59tm7u4489lnahm5fbbjcmplm0.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.911090474857-gjmq3t59tm7u4489lnahm5fbbjcmplm0</string>
+	<key>ANDROID_CLIENT_ID</key>
+	<string>911090474857-gbiv6ervnc8ptjmhtfkvrp8hmv5chmor.apps.googleusercontent.com</string>
+	<key>API_KEY</key>
+	<string>AIzaSyBTjVXj0QISn6PY1gb9yvh6sCzctAH1nFk</string>
+	<key>GCM_SENDER_ID</key>
+	<string>911090474857</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.example.app</string>
+	<key>PROJECT_ID</key>
+	<string>tacklepay-dev</string>
+	<key>STORAGE_BUCKET</key>
+	<string>tacklepay-dev.firebasestorage.app</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:911090474857:ios:97e5ead604b6c580a9e7f8</string>
+</dict>
+</plist>
