+++ app/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:config/config.dart';
 import 'package:core/core.dart';
 import 'package:flutter/material.dart';
