+++ config/lib/config.dart
@@ -1,0 +1,2 @@
+export 'src/modules/app_modules.dart';
+export 'src/router/router.dart';
