+++ config/lib/src/di/configure_dependencies.config.dart
@@ -1,0 +1,29 @@
+// dart format width=80
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// InjectableConfigGenerator
+// **************************************************************************
+
+// ignore_for_file: type=lint
+// coverage:ignore-file
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:auth/util/di/injector.module.dart' as _i872;
+import 'package:core/core.dart' as _i494;
+import 'package:get_it/get_it.dart' as _i174;
+import 'package:home/util/di/injector.module.dart' as _i801;
+import 'package:injectable/injectable.dart' as _i526;
+
+// initializes the registration of main-scope dependencies inside of GetIt
+Future<_i174.GetIt> $initGetIt(
+  _i174.GetIt getIt, {
+  String? environment,
+  _i526.EnvironmentFilter? environmentFilter,
+}) async {
+  final gh = _i526.GetItHelper(getIt, environment, environmentFilter);
+  await _i872.AuthPackageModule().init(gh);
+  await _i801.HomePackageModule().init(gh);
+  await _i494.CorePackageModule().init(gh);
+  return getIt;
+}
