+++ config/lib/src/router/router.dart
@@ -1,0 +1,2 @@
+export 'app_router.dart';
+export 'router_provider.dart';
