+++ config/pubspec_overrides.yaml
@@ -1,6 +1,8 @@
-# melos_managed_dependency_overrides: auth,home
+# melos_managed_dependency_overrides: auth,core,home
 dependency_overrides:
   auth:
     path: ../features/auth
+  core:
+    path: ../core
   home:
     path: ../features/home
