--- core/lib/src/router/app_router.gr.dart
@@ -1,29 +1,0 @@
-// dart format width=80
-// GENERATED CODE - DO NOT MODIFY BY HAND
-
-// **************************************************************************
-// AutoRouterGenerator
-// **************************************************************************
-
-// ignore_for_file: type=lint
-// coverage:ignore-file
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:auto_route/auto_route.dart' as _i2;
-import 'package:core/src/router/wrapper/empty_router_page.dart' as _i1;
-
-/// generated route for
-/// [_i1.EmptyRouterPage]
-class EmptyRouterRoute extends _i2.PageRouteInfo<void> {
-  const EmptyRouterRoute({List<_i2.PageRouteInfo>? children})
-    : super(EmptyRouterRoute.name, initialChildren: children);
-
-  static const String name = 'EmptyRouterRoute';
-
-  static _i2.PageInfo page = _i2.PageInfo(
-    name,
-    builder: (data) {
-      return const _i1.EmptyRouterPage();
-    },
-  );
-}
