--- core/lib/src/router/wrapper/empty_router_page.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../../../core.dart';
-
-@RoutePage()
-class EmptyRouterPage extends StatelessWidget {
-  const EmptyRouterPage({super.key});
-
-  @override
-  Widget build(BuildContext context) => const AutoRouter();
-}
