--- core/lib/src/router/wrapper/wrappers.dart
@@ -1,1 +1,0 @@
-export 'empty_router_page.dart';
