+++ core/lib/src/utils/di/injector.dart
@@ -1,0 +1,6 @@
+// ignore_for_file: inference_failure_on_function_return_type, public_member_api_docs, always_declare_return_types, type_annotate_public_apis, lines_longer_than_80_chars
+
+import '../../../core.dart';
+
+@microPackageInit
+injectorCoreModule() {}
