+++ core/lib/src/utils/di/register_module.dart
@@ -3,9 +3,9 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:logger/logger.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
-import '../../core.dart';
-import '../log/filter/release_log_filter.dart';
-import '../log/printer/pretty_printer.dart';
+import '../../../core.dart';
+import '../../log/filter/release_log_filter.dart';
+import '../../log/printer/pretty_printer.dart';
 
 @module
 abstract class RegisterModule {
