+++ core/lib/src/utils/di/service_locator.dart
@@ -1,0 +1,3 @@
+import 'package:get_it/get_it.dart';
+
+final getIt = GetIt.instance;
