+++ core/lib/src/utils/utils.dart
@@ -1,3 +1,4 @@
 export 'desing_size.dart';
 export 'device_info_util.dart';
+export 'di/injector.module.dart';
 export 'package_info_util.dart';
