+++ features/auth/pubspec.yaml
@@ -14,11 +14,9 @@ dependencies:
   # Internal Packages
   core:
     path: ../../core
-  home:
-    path: ../home
 
   # External Packages
-  firebase_auth: 5.6.1
+  firebase_auth: 6.0.0
   flutter_svg: 2.2.0
   google_sign_in: 6.3.0
 
@@ -30,8 +28,10 @@ dev_dependencies:
   auto_route_generator: 10.2.3
   build_runner: 2.5.4
   injectable_generator: 2.7.0
+
+  # Mock dependencies
   mocktail: 1.0.2
-  firebase_auth_mocks: 0.14.2
+  firebase_auth_mocks: 0.15.0
   google_sign_in_mocks: 0.3.0
 
 flutter:
