+++ melos.yaml
@@ -3,6 +3,7 @@ name: fintech_workspace
 packages:
   - app
   - core
+  - config
   - features/*
 
 command:
@@ -17,8 +18,8 @@ command:
       dartz: 0.10.1
       device_info_plus: 11.5.0
       equatable: 2.0.7
-      firebase_auth: 5.6.1
-      firebase_core: 3.15.0
+      firebase_auth: 6.0.0
+      firebase_core: 4.0.0
       flutter_native_splash: 2.4.4
       flutter_riverpod: 2.6.1
       flutter_screenutil: 5.9.3
@@ -44,7 +45,7 @@ command:
       build_runner: 2.5.4
       injectable_generator: 2.7.0
       mocktail: 1.0.2
-      firebase_auth_mocks: 0.14.2
+      firebase_auth_mocks: 0.15.0
       google_sign_in_mocks: 0.3.0
   clean:
     hooks:
